+++ geopandas/__init__.py
@@ -5,8 +5,8 @@ from geopandas.geodataframe import GeoDataFrame  # noqa
 from geopandas.array import points_from_xy  # noqa
 
 from geopandas.io.file import _read_file as read_file  # noqa
-from geopandas.io.parquet import _read_parquet as read_parquet  # noqa
-from geopandas.io.parquet import _read_feather as read_feather  # noqa
+from geopandas.io.arrow import _read_parquet as read_parquet  # noqa
+from geopandas.io.arrow import _read_feather as read_feather  # noqa
 from geopandas.io.sql import _read_postgis as read_postgis  # noqa
 from geopandas.tools import sjoin  # noqa
 from geopandas.tools import overlay  # noqa
